Cambiado la jerarquia de carpetas para que funcione bien el SDk de facebook (Requiere al menos algo con un ., por convencion com.nombrePaquete)
