+++ app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.mauro.llevame_app_project.LoginActivity">
+    tools:context="com.example.mauro.com.llevame_app_project.LoginActivity">
 
     <!-- Login progress -->
     <ProgressBar
