--- app/src/test/java/llevame_app_project/ExampleUnitTest.java
@@ -1,12 +1,0 @@
-package llevame_app_project;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-
-//Local unit test, which will execute on the development machine (host).
-
-public class ExampleUnitTest {
-
-}
